enabling test suite for el7

+++ tests/p_ipa-server/0-preclean.sh
@@ -2,11 +2,11 @@
 # Author: James Hogarth <james.hogarth@gmail.com>
 #
 
-if (t_GetPkgRel basesystem | grep -q el6)
+if (t_GetPkgRel basesystem | grep -qE 'el(6|7)')
 then
 
 t_Log "Running $0 - Stopping and removing httpd if present"
-if  /sbin/service httpd status 2>&1 | grep 'is running...' &> /dev/null
+if  /sbin/service httpd status 2>&1 | grep 'running...' &> /dev/null
 then
 /sbin/service httpd stop &> /dev/null
 fi
@@ -17,7 +17,7 @@ rm -rf /etc/httpd
 fi
 
 t_Log "Running $0 - Stopping and removing bind if present"
-if  /sbin/service named status 2>&1 | grep 'is running...' &> /dev/null
+if  /sbin/service named status 2>&1 | grep 'running...' &> /dev/null
 then
 /sbin/service named stop &> /dev/null
 fi
@@ -37,8 +37,11 @@ cp /etc/nsswitch.conf /tmp/nsswitch.conf.ipa-tests
 t_Log "Running $0 - Backing up hosts file"
 cp /etc/hosts /tmp/hosts.ipa-tests
 
+if [[ -f /etc/ntp.conf ]]
+then
 t_Log "Running $0 - Backing up ntp.conf file"
 cp /etc/ntp.conf /tmp/ntp.conf.ipa-tests
+fi
 
 t_Log "Running $0 - Backing up saving yum history id"
 /usr/bin/yum history list | awk 'NR == 4 {print $1}' > /tmp/yum-rollback-id.ipa-tests
