+++ tests/p_ipa-server/99-postclean.sh
@@ -2,7 +2,7 @@
 # Author: James Hogarth <james.hogarth@gmail.com>
 #
 
-if (t_GetPkgRel basesystem | grep -q el6)
+if (t_GetPkgRel basesystem | grep -qE 'el(6|7)')
 then
 
 t_Log "Running $0 - Restoring up resolv.conf"
@@ -14,8 +14,11 @@ cp /tmp/nsswitch.conf.ipa-tests /etc/nsswitch.conf
 t_Log "Running $0 - Restoring  hosts file"
 cp /tmp/hosts.ipa-tests /etc/hosts
 
+if [[ -f /tmp/ntp.conf.ipa-tests ]]
+then
 t_Log "Running $0 - Restoring  ntp.conf file"
 cp /tmp/ntp.conf.ipa-tests /etc/ntp.conf
+fi
 
 t_Log "Running $0 - Rolling back to yum history id"
 /usr/bin/yum -y history rollback $(cat /tmp/yum-rollback-id.ipa-tests) &> /dev/null
